// src/domain/interfaces/deposit.interface.ts
import { IAuditable } from './auditable.interface';
import { ITransaction } from './transaction.interface';
import { IUser } from './users.interface';
export interface IDeposit extends IAuditable {
  referenceNo: string;
  amount: number;
  currencyCode: string;
  userId: string;
  user?: IUser;
  narration?: string;
  transactionId: string;
  transaction: ITransaction;
  providerResponse?: string;
}